s2t_system/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.2
billiard==4.2.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorful==0.5.6
cryptography==43.0.0
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-cache2==0.2.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
ffmpeg-python==0.2.0
filelock==3.15.4
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
google-api-core==2.19.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
humanize==4.10.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
linkify-it-py==2.0.3
llvmlite==0.43.0
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.13.4
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numba==0.60.0
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
opencensus==0.11.4
opencensus-context==0.1.3
packaging==24.1
pendulum==3.0.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
pwdlib==0.2.0
py-spy==0.3.14
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.10.1
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
sympy==1.13.1
textual==0.75.1
tiktoken==0.7.0
time-machine==2.15.0
torch==2.3.1
tornado==6.4.1
tqdm==4.66.5
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
vine==5.1.0
virtualenv==20.26.3
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
